+++ src/bare-metal/template/timer.cpp
@@ -5,12 +5,12 @@
 #include "hw_specific.h"
 
 namespace RODOS {
-long long Timer::microsecondsInterval = 0;
+int64_t Timer::microsecondsInterval = 0;
 
 void Timer::init()  { }
 void Timer::start() { }
 void Timer::stop()  { }
-void Timer::setInterval([[gnu::unused]] const long long microsecondsInterval) { }
+void Timer::setInterval([[gnu::unused]] const int64_t microsecondsInterval) { }
 
 
 }
