+++ src/on-posix/hw_specific.h
@@ -59,7 +59,7 @@ extern int32_t getNodeNumber();
 class Timer {
 
   private:
-    static long long microsecondsInterval;
+    static int64_t microsecondsInterval;
     Timer();
 
   public:
@@ -83,12 +83,12 @@ class Timer {
     /**
   *  Set timer interval.
   */
-    static void setInterval(const long long microsecondsInterval);
+    static void setInterval(const int64_t microsecondsInterval);
 
     /**
   *  Get timer interval.
   */
-    static long long getInterval() { return microsecondsInterval; }
+    static int64_t getInterval() { return microsecondsInterval; }
 };
